#  Optimize GEORADIUS command performance with  pre-allocated buffer