--- a/src/geo.h
+++ b/src/geo.h
@@ -3,6 +3,8 @@
 
 #include "server.h"
 
+#define MAX_GEO_ARRAY_BUFFER 8
+
 /* Structures used inside geo.c in order to represent points and array of
  * points on the earth. */
 typedef struct geoPoint {
@@ -17,6 +19,7 @@ typedef struct geoArray {
     struct geoPoint *array;
     size_t buckets;
     size_t used;
+    struct geoPoint arraybuf[MAX_GEO_ARRAY_BUFFER]; /* Pre-allocated buffer reduces heap allocation */
 } geoArray;
 
 #endif