**Description**

1. The GEORADIUS command stores searched map points in geoArray, which triggers at least two memory allocations: one for the `geoArray` creation and another for each `geoPoint` initialization.
2. We can reduce these two memory allocations: 
   - converting `geoArray` from a heap-allocated variable to a stack-allocated one
   - embedding `geoPoint` directly within geoArray
  
**Benchmark**
Run the benchmark command five times with the command below, and take the peak value as the final result.
> for((x=0;x<5;x+=1));
>do
>valkey-cli flushall;
>valkey-benchmark -r $POINTS_SIZE geoadd key 100 70 l__rand_int__;
>valkey-benchmark -P 2 -n 10000000 georadius key 100 70 300 km;
> done
> 

$POINTS_SIZE | QPS before optimization|QPS after optimization|Performance Boost
-- | -- | -- | -- |
1|430015.06|436585.88|1.5%|
4|381766.81|386279.34|1.2%|
16|264592.28|268391.53|1.4%|
64|120035.05|120277.60|0.2%|

CPU: AMD EPYC 9K65 192-Core Processor* 8
OS: 	Ubuntu Server 24.04 LTS 64bit
Memory: 32GB
VM: Tencent cloud  SA9 | SA9.2XLARGE32